Fix warnings when building as x64.
Caused by updates in: d834c128bb60bdcc910099725a717d195e8598ba

+++ Pulp/Font/XFont.cpp
@@ -25,7 +25,7 @@ namespace
 	{
 		XFont* pFont = static_cast<XFont*>(gEnv->pFont);
 
-		int Num = pCmd->GetArgCount();
+		size_t Num = pCmd->GetArgCount();
 
 		if (Num < 2)
 		{
